Refactoring static files
