+++ spring-boot-electron/stockapp/package.json
@@ -4,14 +4,14 @@
   "description": "Prices of commodities",
   "license": "MIT",
   "repository": "spring-boot-electron/stockapp",
-  "main": "src/main/js/app.js",
+  "main": "src/main/resources/static/js/app.js",
   "dependencies": {
     "papaparse": "^4.6.0",
     "rest": "^1.3.1"
   },
   "scripts": {
     "watch": "webpack --watch -d",
-    "start": "electron src/main/js/app.js"
+    "start": "electron src/main/resources/static/js/app.js"
   },
   "devDependencies": {
     "electron": "^3.0.2",
