+++ spring-boot-electron/stockapp/src/main/resources/static/css/styles.css
